requirements.txt
crewai
crewai_tools
openai
python-dotenv
langchain
langchain-core
pydantic